experimento-autorizar-actores/autorizador/requirements.txt
flask==3.0.3
pyjwt[crypto]==2.9.0
requests==2.32.3
python-dotenv==1.0.1
gunicorn==21.*